+++ packages/utils/package.json
@@ -4,6 +4,7 @@
   "type": "module",
   "exports": {
     "./belt-level": "./src/belt-level.ts",
+    "./download": "./src/download.ts",
     "./cache": "./src/cache.ts",
     "./categories": "./src/categories.ts",
     "./constants": "./src/constants.ts",
