+++ packages/utils/src/download.ts
@@ -1,0 +1,15 @@
+export function downloadFile(content: string, filename: string, type: string): void {
+  const blob = new Blob([content], { type });
+  const url = URL.createObjectURL(blob);
+  const a = document.createElement("a");
+  a.href = url;
+  a.download = filename;
+  document.body.append(a);
+  a.click();
+  a.remove();
+  URL.revokeObjectURL(url);
+}
+
+export function downloadJson(data: object, filename: string): void {
+  downloadFile(JSON.stringify(data, null, 2), filename, "application/json");
+}
